added application customization test
