+++ lib/views/main_view/main_view_widgets/token_widgets/default_token_actions/default_lock_action.dart
@@ -4,7 +4,7 @@ import 'package:flutter_slidable/flutter_slidable.dart';
 import '../../../../../l10n/app_localizations.dart';
 import '../../../../../model/enums/introduction.dart';
 import '../../../../../model/tokens/token.dart';
-import '../../../../../utils/app_customizer.dart';
+import '../../../../../utils/customization/action_theme.dart';
 import '../../../../../utils/lock_auth.dart';
 import '../../../../../utils/logger.dart';
 import '../../../../../utils/riverpod_providers.dart';
