+++ lib/views/splash_screen/splash_screen.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
 import '../../model/enums/app_feature.dart';
-import '../../utils/app_customizer.dart';
+import '../../utils/customization/application_customization.dart';
 import '../../utils/app_info_utils.dart';
 import '../../utils/home_widget_utils.dart';
 import '../../utils/logger.dart';
