+++ lib/widgets/home_widgets/home_widget_otp.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 
-import '../../utils/app_customizer.dart';
+import '../../utils/customization/extended_text_theme.dart';
 import '../../utils/utils.dart';
 import 'interfaces/flutter_home_widget_base.dart';
 import 'interfaces/flutter_home_widget_builder.dart';
