--- test/unit_test/utils/app_customizer_test.dart
@@ -1,1 +1,0 @@
-
